moved some files , added some mock data for books, updated the logic of user list

+++ repository/user_list.interface.go
@@ -4,7 +4,7 @@ import "github.com/kdsama/book_five/domain"
 
 type UserListRepo interface {
 	SaveUserList(*domain.UserList) error
-	CountExistingListsOfAUser(user_id string) (int, error)
+	CountExistingListsOfAUser(user_id string) (int64, error)
 }
 
 type UserListRepository struct {
