+++ service/user_list.interface.go
@@ -2,7 +2,7 @@ package service
 
 type UserListServiceInterface interface {
 	SaveUserList(user_id string, about string, list_name string, book_ids []string) error
-	CountExistingListsOfAUser(user_id string) (int, error)
+	CountExistingListsOfAUser(user_id string) (int64, error)
 }
 
 type UserListDI struct {
